Eliminate Bottlenecks - LLM assistant/the_demo_app/requirements.txt
streamlit>=1.36
llama-index>=0.10.0
llama-index-readers-file>=0.1.0
llama-index-llms-openai>=0.1.0
openai>=1.0.0
fastmcp>=0.4.0
python-dotenv>=1.0.1
uvicorn>=0.30.0
sqlalchemy>=2.0.0


